+++ vite.config.ts
@@ -2,6 +2,8 @@ import {defineConfig, loadEnv} from 'vite'
 import react from '@vitejs/plugin-react-swc'
 import defineCustomConfig from './define.config';
 import storageRPostBuild from "./post.build";
+import { version } from './package.json';
+
 
 export default defineConfig(({mode}) => {
   const env = loadEnv(mode, process.cwd());
@@ -13,7 +15,14 @@ export default defineConfig(({mode}) => {
       outDir: outDir,
       emptyOutDir: true,
       sourcemap: mode === 'test',
-      minify: mode === 'test' || mode === 'development' ? false : 'esbuild'
+      minify: mode === 'test' || mode === 'development' ? false : 'esbuild',
+      rollupOptions: {
+        output: {
+          assetFileNames: `assets/[name]-[hash]-v${version}[extname]`,
+          chunkFileNames: `chunks/[name]-[hash]-v${version}.js`,
+          entryFileNames: `[name]-[hash]-v${version}.js`,
+        },
+      },
     },
     plugins: [react(), storageRPostBuild({outDir: outDir})],
     define: defineCustomConfig.defineViteConfig,